backend/pyproject.toml
[project]
name = "test-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "mysql-connector-python>=9.5.0",
    "redis>=7.1.0",
    "flask-session>=0.8.0",
    "flask-limiter>=3.5.0",
]
